Rename activites app to actions
