+++ apps/actions/templates/meinberlin_actions/blocks/platform_activity_block.html
