--- apps/activities/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class Config(AppConfig):
-    name = 'apps.activities'
-    label = 'meinberlin_activities'
